test: extract crud buttons tests
