+++ packages/crud/test/crud-editor.test.js
@@ -3,31 +3,62 @@ import { fixtureSync, nextRender } from '@vaadin/testing-helpers';
 import '../vaadin-crud.js';
 import { flushGrid } from './helpers.js';
 
-describe('vaadin-crud editor', () => {
-  ['default', 'custom'].forEach((type) => {
-    const tpl =
-      type === 'default'
-        ? '<vaadin-crud style="width: 300px;"></vaadin-crud>'
-        : `<vaadin-crud style="width: 300px;">
-        <vaadin-form-layout slot="form">
-          <vaadin-text-field path="foo" required></vaadin-text-field>
-        </vaadin-form-layout>
-      </vaadin-crud>`;
+describe('crud editor', () => {
+  let crud;
 
+  describe('header', () => {
+    let header;
+
+    beforeEach(async () => {
+      crud = fixtureSync('<vaadin-crud style="width: 300px;"></vaadin-crud>');
+      crud.items = [{ foo: 'bar' }];
+      header = crud.querySelector('[slot=header]');
+      await nextRender(crud._grid);
+    });
+
+    it('should have new item title', () => {
+      crud.$.new.click();
+      expect(header.textContent).to.be.equal('New item');
+    });
+
+    it('should have edit item title', () => {
+      crud.editedItem = crud.items[0];
+      expect(header.textContent).to.be.equal('Edit item');
+    });
+
+    it('should change to new item title', () => {
+      crud.editedItem = crud.items[0];
+      expect(header.textContent).to.be.equal('Edit item');
+      crud.$.new.click();
+      expect(header.textContent).to.be.equal('New item');
+    });
+  });
+
+  ['default', 'custom'].forEach((type) => {
     describe(`${type} form`, () => {
-      let crud, dialog, form, btnCancel, overlay;
+      let dialog, form, btnCancel, overlay;
 
       beforeEach(async () => {
-        crud = fixtureSync(tpl);
+        if (type === 'default') {
+          crud = fixtureSync('<vaadin-crud style="width: 300px;"></vaadin-crud>');
+        } else {
+          crud = fixtureSync(`
+            <vaadin-crud style="width: 300px;">
+              <vaadin-form-layout slot="form">
+                <vaadin-text-field path="foo" required></vaadin-text-field>
+              </vaadin-form-layout>
+            </vaadin-crud>
+          `);
+        }
         crud.editOnClick = true;
         crud._fullscreen = false;
         await nextRender(crud);
         flushGrid(crud._grid);
 
         crud.set('items', [{ foo: 'bar' }, { foo: 'baz' }]);
-        form = crud._form;
         dialog = crud.$.dialog;
         overlay = dialog.$.overlay;
+        form = crud.querySelector('[slot=form]');
         btnCancel = crud.querySelector('[slot="cancel-button"]');
       });
